Add a script to bootstrap running Pester tests (#3899)
Adds a script that can bootstrap from nothing to run the Pester tests.
Also updates the existing test scripts to enable this with the goal of
not changing their current behavior. Publishes the test scripts so that
one can download the artifacts and run tests against whatever set of
modules and client is interesting.