Allow user settings to control logging channels (#3955)
Add a user setting to control the enabled logging channels, then disable
the `SQL` by default. This channel is responsible for 95% of the log
lines in a verbose file, yet they are only helpful when debugging a very
specific set of issues that does not occur often.

Additionally, make the COM logging use the settings for logging (both
channels and level).