Introduce strong and weak comparisons between installers (#3956)
Rather than a `bool`, the installer comparison function now returns one
of three values: negative, weak positive, and strong positive. The
comparison functions are updated to choose between weak and strong
positive results as appropriate. For instance, matching the system
architecture when the alternative does not is a strong positive result.
Simply being "better" in the list of emulated architectures is a weak
positive result.

The function that used these comparators in our priority order is
updated to effectively do multiple passes, one looking for strong
results and then one looking for weak results (in reality it is
implemented with one pass as an optimization).