Cherry-Pick from #3937 - Fix pipeline build error (#4063)
This fixes the recent build errors in the pipelines. The compiler was
complaining about re-throwing an exception with `throw e;` instead of
`throw;`, which messes with the stack trace. Presumably it started
because of a tool update in the agents.

This also fixes a compilation error I'm seeing locally from a member not
being initialized.

###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/microsoft/winget-cli/pull/4063)