Use std::variant in ManifestYamlPopulator (#4081)
This PR uses `std::variant` instead of member pointers to manifest
structs to be references used by the processing functions.
`std::variant` contains pointers to manifest structs to be populated and
is passed to the `FieldProcessInfo.ProcessFunc`. Each processing
function is responsible of calling `std::get` or `std::get_if` with the
correct pointer type. `ValidateAndProcessFields` takes the
`std::variant` and pass it to the processing functions.

Processing functions that required the member types now take a pointer
to the manifest struct. Some of these can go in the anonymous namespace,
but I didn't want to add more moves.

The change will make implementing shadow manifests easier.