Support group processing of configurations (#4059)
Adds support for test and apply processing to handle groups. This is
done by changing the top-level processing to only work on groups, and
moving the existing processing code to a default group processor that is
used when the set processor does not expose the group interface. All
telemetry is handled by the top level processing code so that we get
consistent data regardless of the set processor.

Both set and unit processors can implement
`IConfigurationGroupProcessor` to indicate that they can handle group
processing on their own. If the set processor does not expose this
interface, the default processor will handle the set level units while
checking the unit processors for group handling as well. If the set
processor does expose the interface, it is expected to handle all
processing. It is still necessary to implement unit processors to expose
details about them, but they don't need to be able to function.
  
Also changed most `winrt::single_threaded_vector` to
`winrt::multi_threaded_vector` as it was appropriate for their potential
usage.