Improve progress handling for group processor (#4121)
To prevent the loss of progress events between the group processor and
the caller, move to a model where these internal interfaces take in an
`EventHandler` parameter rather than using `...WithProgress`. This will
make it possible to be confident that no progress events occurred
between the async method starting and the progress handler being
attached.

This fixes the flakiness of the tests that were verifying progress
events.