Shadow Manifest (#4104)
This PR introduces shadow manifests.

A shadow manifest is a file used to create a merged manifest that
contains extra information about the package. For now, the shadow
manifest contains icon information. The file is meant to be created
programmatically and can be done via `ManifestShadow` from the
`WinGetUtilnterop` nuget package.

The shadow will be added to the merged manifest when a directory
contains a shadow manifest (ManifestType = shadow) and the create
options allow using shadow manifest. If the flag is not set and a shadow
manifest file is found it will fail. There can only be one shadow
manifest file in the directory.

The shadow manifest gets processed and merged into the `Manifest` and
the root YAML node. It will only merge information if the processed
merged manifest doesn't already contain it. For example, if the manifest
already contains a localization node with a specific `PackageLocale`
that have `Icons` and the shadow manifest also contains a localization
node with the same `PackageLocale` value with `Icons` it won't get
merged.

A serialized shadow manifest looks like this (ignore values):
```
PackageIdentifier: Package.package
PackageVersion: 1.0
ManifestType: shadow
ManifestVersion: 1.5
PackageLocale: en-US
Icons:
- IconUrl: iconUrl
  IconFileType: fileType
  IconResolution: iconResolution
  IconTheme: iconTheme
  IconSha256: iconSha256
Localization:
- PackageLocale: es-MX
  Icons:
  - IconUrl: iconUrl-esMX
    IconFileType: fileType-esMX
    IconResolution: iconResolution-esMX
    IconTheme: iconTheme-esMX
    IconSha256: iconSha256-esMX
- PackageLocale: de-DE
  Icons:
  - IconUrl: iconUrl-de-DE
    IconFileType: fileType-de-DE
    IconResolution: iconResolution-de-DE
    IconTheme: iconTheme-de-DE
    IconSha256: iconSha256-de-DE
```
###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/microsoft/winget-cli/pull/4104)