Refactor `IPackage` (#4174)
## Change
To better reflect the underlying code and continue to move business
logic out of the `CompositeSource`, this change refactors the `IPackage`
interface into 3:

```C++
    // Contains a collection of package versions.
    struct IPackageVersionCollection
    {
        // Gets all versions of this package.
        // The versions will be returned in sorted, descending order.
        //  Ex. { 4, 3, 2, 1 }
        virtual std::vector<PackageVersionKey> GetVersionKeys() const = 0;

        // Gets a specific version of this package.
        virtual std::shared_ptr<IPackageVersion> GetVersion(const PackageVersionKey& versionKey) const = 0;

        // A convenience method to effectively call `GetVersion(GetVersionKeys[0])`.
        virtual std::shared_ptr<IPackageVersion> GetLatestVersion() const = 0;
    };

    // Contains information about a package and its versions from a single source.
    struct IPackage : public IPackageVersionCollection
    {
        // Gets a property of this package.
        virtual Utility::LocIndString GetProperty(PackageProperty property) const = 0;

        // Gets the source that this package is from.
        virtual Source GetSource() const = 0;

        // Determines if the given IPackage refers to the same package as this one.
        virtual bool IsSame(const IPackage*) const = 0;

        // Gets this object as the requested type, or null if it is not the requested type.
        virtual const void* CastTo(IPackageType type) const = 0;
    };

    // Contains information about the graph of packages related to a search.
    struct ICompositePackage
    {
        // Gets a property of this package result.
        virtual Utility::LocIndString GetProperty(PackageProperty property) const = 0;

        // Gets the installed package information.
        virtual std::shared_ptr<IPackage> GetInstalled() = 0;

        // Gets all of the available packages for this result.
        // There will be at most one package per source in this list.
        virtual std::vector<std::shared_ptr<IPackage>> GetAvailable() = 0;
    };
```

`IPackageVersionCollection` is simply a set of `IPackageVersion` objects
with only the previous `IPackage` functions to enumerate them. This
enables some code that only wants to inspect versions to use a tighter
set of APIs and thus some implementations need not worry about things
like the appropriate property values to return.

The refactored `IPackage` inherits from `IPackageVersionCollection` and
is itself a set of package versions that all come from a single source
(as in `Source`). It adds the properties and the `Source` that the
versions are all from.

`ICompositePackage` is the new search result, resembling the previous
`IPackage` in what it contains but exposing it more directly. The
installed `IPackage` object will eventually be able to contain more than
one installed version in a future change. The available packages are
exposed directly as `IPackage` objects, rather than mixing the versions
together in the `CompositeSource`.

The code in `PackageVersionSelection.cpp` extracts the business logic
that was inside `CompositeSource` previously, as well as providing a few
helpers to make the existing code work more succinctly. The refactoring
is not attempting to change behavior, but simply make it possible to
change behavior more easily and precisely in the future.

I would have preferred to go further by splitting `CompositeSource` out
into a public type, allowing the other sources to have a simpler model
of only ever returning `IPackage` results from search. This change would
have created an even larger churn though and did not directly contribute
towards the goal of supporting future side-by-side work.
  
Also changes the YAML object model to use a non-exceptional function to
attempt to convert to an integer.