Better server start coordination (#4292)
## Change
Improve this named event creation by having both sides try to create or
open it. The current version will almost always have to sleep at least
once while the server sets up and can completely fail if it takes too
long. With this change, we should be able to properly wait for startup
on the first attempt.