Add Workflow logs and fix installed version workflow bug (#4432)
Fixes #4425 

## Change
This adds a `Workflow` channel (disabled by default) that has the entry
to tasks and the `Get`/`Add`/`Contains` calls logged to it. This enables
a view into a flow that can be collected by a third party.

It also fixes a behavior change with the side-by-side code path for
selecting the installed version that would lead to no data item being
inserted, when the previous behavior was that a `nullptr` would be
inserted when no version is installed.