Configure export command (#4434)
This PR introduces the `configure export` command as an exprimental
feature. This is mostly a proof of concept and should not be considered
a full feature.

#### Scenario A - Create a configuration to install a winget package.
Use `--pacakgeId` with the package identifier of an application in the
winget source to produce a configuration file that uses the
`Microsoft.WinGet.DSC/WinGetPackage` resource to install the package via
winget. Right now, we don't validate if the package id exists and will
just copy what the user sets into the settings of the resource.

#### Scenario B - 'Export' the configuration from the specified
resource.
Use both `--module` and `--resource` to get the configuration of a
resource and add it to the configuration file. Internally, configuration
will install the module (if not installed already) and call Get on the
resource. We will then try to serialize its property into the
configuration file. Current limitation is that if a resource has a
required setting (like WinGetPackage required Id) it won't work. If the
resource is not found in the gallery a retry will be performed allowing
prereleased modules in the case it exists.

#### Scenario C - Mix of A and B
If `--packageId`, `--module` and `--resource` are used, configure export
will produce two resources. The first one is the `WinGetPackage` for the
specified package. The second one is the same as in B, with the
difference that it includes a dependency of the previously created
`WinGetPackage` resource.

#### Scenario D - Configuration file already exists.
If the file passed to the `--output` parameters already exists and is a
valid configuration file, the resources will be appended. There is
currently not validation into the correctness of this, so it can result
in a configuration with resources with the same id.

For example `winget configure export --packageId Microsoft.AppInstaller
--module Microsoft.WinGet.DSC --resource WinGetUserSettings -o
test_export.yml` would produce the following file

```
# Created using winget configure export 1.8.0-preview
# yaml-language-server: $schema=https://aka.ms/configuration-dsc-schema/0.2
properties:
  configurationVersion: 0.2
  resources:
  - resource: Microsoft.WinGet.DSC/WinGetPackage
    id: Microsoft.AppInstaller
    directives:
      description: Install Microsoft.AppInstaller
      allowPrerelease: true
    settings:
      id: Microsoft.AppInstaller
      source: winget
  - resource: Microsoft.WinGet.DSC/WinGetUserSettings
    dependsOn:
    - Microsoft.AppInstaller
    directives:
      description: Configure Microsoft.AppInstaller
    settings:
      Settings:
        experimentalFeatures:
          configureSelfElevate: true
        installBehavior:
          preferences:
            locale:
            - en-US
            - fr-FR
        $schema: https://aka.ms/winget-settings.schema.json
      Action: Full
```


###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/microsoft/winget-cli/pull/4434)