Configuration apply queueing (#4590)
## Change
Adds a queue table to the configuration database and some code to
synchronize the application of configurations.

Every apply in the queue puts a row in the table, with its instance
identifier (it should also be in the history) and a named object that it
will keep alive as long as it is in the queue. This allows for other
queued processes to check for dead queue items.

A global named mutex must be held in order to apply, or even check if
one is at the front of the queue. If not at the front of the queue, the
waiting operation will release the mutex and wait for N * 100ms where N
is their perceived position in the queue. This should prevent repeated
contention on the global mutex as the queued items sort themselves via
the wait.