Delete CodeCoverage and WinGetUtil NuGet pipelines (#4708)
The pipeline for fuzzing code coverage is no longer needed.

The pipeline for creating the wingetutil nuget will be merged with the
internal release pipeline that produces all other release artifacts, and
nuspec file will also be added to internal repo alongside the other
nuspec files.

This only removes the pipeline definitions. Pipelines in AzDO will be
deleted later.