Create public sample (#4734)
## Change
Copy the current dev tool, COM calling application into a new location
and update significantly with the goal of being a public sample of
usage. Cleaned up the code and tried to include more error handling and
commenting.

Major changes to the UX:
- Removed the dev switch as this is intended as a public sample
- Added example search options selection
- Add `Upgrade` button so that function could be called as well
- Add `Uninstall` button for installed packages
- Disconnected installed and active operation lists from catalog
selection and package search

Major change to DevX:
- Move to using the OOP COM nuget package
- This isn't ready for use yet, as the OOP activation dll isn't
available in a public version yet and we are still working on the exact
shape of that update.