Fix OOP COM Repair Test Failures
The following tests failed in the Azure pipeline due to ERROR_ADMIN_CONTEXT_REPAIR_PROHIBITED (0x8A15007D):
- DisableWinGetCommandLineInterfacesPolicy
- RepairBurnInstallerWithModifyBehavior
- RepairExeInstallerWithUninstallerBehavior
- RepairNullsoftInstallerWithUninstallerBehavior

The problem is due to the OOP COM Server running as admin in the pipeline, which is not the case for local test runs. This discrepancy may indicate different test configurations in the pipeline. This prevents the repair of user-scope installed packages. The solution is to install the test package to system scope, similar to In-Proc tests.

-  Made minor optimizations to the replacement arguments string building logic.