Remove preference for "installed" installer type (#4740)
Fixes #2565 

## Change
Move (and rename) the `InstalledTypeComparator` from being a comparator
to just a filter. Given what we can (or rather, currently do) actually
discern from system state, it doesn't make sense to give preference to
the few types that we use to signify the compatibility sets. If we truly
want to preserve installer type selection within a compatibility set, we
need to store the user selection into the tracking database similar to
`UserIntentArchitecture`.