Add hash mismatch telemetry details (#4857)
## Issue
From looking at our hash mismatch data, there were two main modes:

1. Lots of actual hashes that are the result of hashing a zero-byte file
2. Other hashes with small numbers of cases, including some non-one
counts. This suggests at least some cases with consistent results even
if they are not what we expected.

My suspicion is that there are some cases where a successful HTTP
response is made, but the content is actually a webpage or similar. This
would happen with captive portals or potentially firewalls.

## Change
Collect the total size and content type of downloads and add them to our
existing hash mismatch telemetry events.
Retry downloading when we get a zero-byte file.
Report a different error when a zero-byte file is the final result of
downloading (also requires the hash to not match, so you can still have
a zero-byte installer for what that is worth...).

## Validation
Updates the downloader tests for new fields.
Adds a new pair of tests for zero-byte file downloads.