Consume all input during extract sequence (#4882)
## Change
When extracting a sequence, consume all (up to 1KB really) input even if
it doesn't start with an escape character.

This was spurred by an issue where the `peek` call was not returning
(waiting for input). Pressing Enter broke it out of that wait, but other
input was now at the front and blocked future attempts to extract the
sequence.

The best solution to this would be to move all input handling to its own
dedicated thread, but that is a much larger change.