Change meaning of AddOrUpdate return bool (#4885)
## Issue
The requestor of this new (1.10) API needs to know whether the operation
was an add or an update. At the same time, the "was the index modified"
return value from Update has always been true for any changes since the
manifest hash was added.

## Change
Return `true` when adding and `false` when updating.